BBB-Back/requirements.txt
asgiref==3.5.2
black==22.10.0
click==8.1.3
Django==4.1.3
django-admin-interface==0.23.0
django-colorfield==0.8.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
djangorestframework==3.14.0
mypy-extensions==0.4.3
pathspec==0.10.2
Pillow==9.3.0
platformdirs==2.5.4
python-http-client==3.3.7
pytz==2022.6
sendgrid==6.9.7
six==1.16.0
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
tomli==2.0.1
